Fixes a couple of bugs in findAltStart. Main.pl now imports subroutines correctly.
One bug caused warning messages to appear when translating genes that had internal stop codons.
This was caused by bad stop codon analysis in the findAltStart algorithm. As a result, the portion
of the code responsible for this analysis has been remade into its own subroutine and improved.

Main.pl used to incorrectly import a couple of subroutines (it looked in the wrong modules for them).
This has also been fixed.