findAltStart remade with new test, reorganized test data
findAltStart.pm has been remade from the ground up using a different
algorithm idea. This time it is not pseudo-recursive and one should
hopefully no longer need to spend an entire afternoon trying to
remember how it was meant to work.

findAltStart.t now has an added test to see if findAltStart.pm is
capable of translating a gene with a known suitable extension. The
previous test to see if a correct gene/protein had been retrieved has
been corrected, so that it no longer compares an object with a string.
The test program has also been cleaned up a little, removing some
obsolete comments and no longer used variables.

Maproutines.pm now has an added condition to handle errors where the
sought genome key (species ID) is not found in genomeHash (likely
occurs because the species IDs used as keys in the hash are named
according to a different naming convention).

Main.pl now has a line that can be uncommented/commented if one wishes
to find the number of processor cores on a Mac system (for
multithreading purposes).

Test genomes have been moved around to appropriate test folders.
Relevant changes has been made for hard coded paths to these in the
affected genome files.